+++ AssignmentHb/src/main/java/service/CustomerService.java
@@ -7,11 +7,11 @@ import entity.Customer;
 public interface CustomerService {
 	void save(Customer customer);
 
-	Customer get(int id);
+	Customer get(long id);
 
 	void update(Customer customer);
 
-	void delete(int id);
+	void delete(long id);
 
 	List<Customer> getAll();
 }
