+++ AssignmentHb/src/main/java/service/OrderService.java
@@ -7,12 +7,12 @@ import entity.Order;
 public interface OrderService {
 	void save(Order order);
 
-	Order get(int id);
+	Order get(long id);
 
 	List<Order> getAll();
 
 	void update(Order order);
 
-	void delete(int id);
+	void delete(long id);
 
 }
